Add general auth test
This test is simply to verify that the Authentication plugin is enabled
on all routes not explicitly allowlisted to be public, when starting the
application using the application module function.
Any explicit handling of the authentication(e.g. access via
roles/id/team etc) is left to test for each route.