Fix history cleanup SQL
We introduced a bug that broke the cleanup function. 'INTERVAL X DAYS'
is a literal in SQL, so trying to inject the number of days via
statement parameters did not work. Fixed it by using the make_interval
function available in Postgres.

Also refactored the dump function to remove the mutable list.