Add stronger type for UserID and simplify auth
Discovered that a few of the auth functions could be made redundant by
moving the ktor-userid-stuff to the route layers. To ensure that the
correct type was sent, the value class UserID was added.
I also colocated config related to Auth with the rest of the config, so
that we now have one place where the config is read and created.